+++ tests/infeed_test.py
@@ -19,7 +19,6 @@ from unittest import SkipTest
 from absl.testing import absltest
 import jax
 from jax import lax, numpy as jnp
-from jax.experimental import host_callback as hcb
 from jax._src import core
 from jax._src import xla_bridge
 from jax._src.lib import xla_client
@@ -77,7 +76,6 @@ class InfeedTest(jtu.JaxTestCase):
 
   @jax.numpy_rank_promotion("allow")  # Test explicitly exercises implicit rank promotion.
   def testInfeedThenOutfeed(self):
-    hcb._deprecated_stop_outfeed_receiver()
 
     @jax.jit
     def f(x):
@@ -99,7 +97,6 @@ class InfeedTest(jtu.JaxTestCase):
     self.assertAllClose(out, y + np.float32(1))
 
   def testInfeedThenOutfeedInALoop(self):
-    hcb._deprecated_stop_outfeed_receiver()
 
     def doubler(_, token):
       y, token = lax.infeed(